# src/main/java/br/com/sgm/iam/model/UserModel.java
package br.com.sgm.iam.model;

import lombok.Builder;
import lombok.Data;

import java.util.ArrayList;
import java.util.List;


@Data
@Builder
public class UserModel  {

    private String id;
    private String name;
    private String email;
    private String password;
    private String username;

    private List<String> profiles = new ArrayList<>();

}